change pipeline default behavior for documentation publishing
this changes the Jenkins pipeline to restrict documentation publishing
on a per-pipeline level.

It also changes pipeline defaults so that
Jenkinsfiles in a default setting are even simpler to write.

It also fixes a bug where documentation would be published for every
branch that was not also a pull request which was not intentional.