fix parameter parsing
previously the conversion from double to string caused problems
if a rather small or rather large number is converted, the scientific
notation is used
e.g. 0.000001 is outputted as 1e-6. yaml 1.1 (which was used in loading
parameters) requires the decimal point (i.e. 1.0e-6)
this change avoids the issue by avoiding string conversion