fix(team_comm): switch to RobotArray msgs on `robots_relative`
topic, instead of ObstaclRelativeArray on `obstacles_relative` topic,
where we do not publish any msgs.

Currently the vision produces RobotArray msgs on `robots_in_image`,
which are then transformed relative to the `base_footprint` map frame to
have relative coordinates based on the running robot's location.
This is done in the ipm (inverse projection mapping), which in turn
publishes `robots_relative`.
For now we totally humans and other obstacles, which could be handled
with ObstaclRelative msgs.

Currently the TeamData msgs also still contains an ObstaclRelativeArray,
because change it would most likely also require changes in the behavior
module.
This will be done at a later stage.

Fixes: https://github.com/bit-bots/humanoid_league_misc/issues/125